requirements.txt
# Prod
web3>=6.8.0
requests>=2.31.0
python-dotenv==0.21.0

# Dev
types-requests>=2.31.0
black==22.10.0
mypy==0.982
pylint==2.15.5
pytest==7.2.0
